# BarToYou-WebApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarToYou API",
        "description": "Documentación Swagger de la API de BarToYou",
        "contact": {
            "email": "soporte@bartoyou.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bartoyou/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Listar pedidos",
                "operationId": "651029e5d3166a1658c4429b585b298c",
                "responses": {
                    "200": {
                        "description": "Lista paginada de pedidos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear un nuevo pedido",
                "operationId": "be4bec2a3ca8406710f01b439fc9074a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "member_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "consumption_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "date_time": {
                                        "type": "string",
                                        "example": "2025-05-08T12:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado correctamente"
                    },
                    "400": {
                        "description": "El pedido ya existe"
                    }
                }
            }
        },
        "/api/bartoyou/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener pedido por ID",
                "operationId": "ca3833626aec34cb2e0a466c21cf7453",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado"
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Actualizar un pedido",
                "operationId": "dc9ba500a985038f878a830cdb091f69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "date_time": {
                                        "type": "string",
                                        "example": "2025-05-09T13:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido actualizado correctamente"
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Eliminar un pedido",
                "operationId": "f347269123fd0305ddb7139bedc50f03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido eliminado correctamente"
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    }
                }
            }
        },
        "/api/bartoyou/orders/user/{userId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener pedidos por usuario",
                "operationId": "67215369ded44133608a638b3c8693d2",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pedidos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "orderid": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "custom_drink_id": {
                                                "type": "string",
                                                "example": "#1234"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "date_time": {
                                                "type": "string",
                                                "example": "2025-05-08T12:30:00"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "En preparación"
                                            },
                                            "items": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Bebida Personalizada"
                                                        },
                                                        "ingredients": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "ingredient": {
                                                                        "type": "string",
                                                                        "example": "Azúcar"
                                                                    },
                                                                    "amount": {
                                                                        "type": "string",
                                                                        "example": "1 cucharadita"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron pedidos para el usuario"
                    }
                }
            }
        },
...
    },
...
}